+++ R/ggbertinplot.R
@@ -19,16 +19,136 @@
 
 ggbertinplot <- function(x,
   order = NULL,
+  highlight = TRUE,
   labRow = TRUE,
   labCol = TRUE,
-  flip = TRUE,
-  ...) {
+  reverse = FALSE,
+  prop = FALSE,
+  geom = "bar") {
+  check_installed("ggplot2")
+
   if (!is.matrix(x))
     stop("Argument 'x' must be a matrix.")
 
-  check_installed("ggplot2")
+  if (is.logical(highlight) && highlight)
+    highlight <- mean(x, na.rm = TRUE)
 
-  ggpimage(x, order, labRow = labRow, labCol = labCol, flip = flip, ...) +
-    ggplot2::scale_x_discrete(position = "top") +
-    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0, vjust = .5))
+  # TODO: rename to circle, box etc.
+  # Add square
+  geom <-
+    match.arg(tolower(geom),
+      choices = c("raster", "block", "circle", "line", "bar", "none"))
+
+  # reorder
+  if (!is.null(order))
+    x <- permute(x, order)
+
+  # change x and y?
+  if (!reverse) {
+    x <- t(x)
+    tmp <- labRow
+    labRow <- labCol
+    labCol <- tmp
+  }
+
+  g <-
+    .ggpimage_empty(
+      x,
+      labRow = labRow,
+      labCol = labCol,
+      prop = prop,
+      expand = geom != "raster"
+    )
+
+  # put col labels on top (message about replacing scale for x)
+  suppressMessages(
+    g <- g +
+      ggplot2::scale_x_discrete(position = "top", expand = if (geom != "raster") ggplot2::waiver() else c(0, 0)) +
+      ggplot2::scale_y_discrete(position = "right", expand = if (geom != "raster") ggplot2::waiver() else c(0, 0)) +
+      ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0, vjust = .5)) +
+      ggplot2::theme(legend.position="bottom")
+  )
+
+  # add geom
+
+  # raster does not use highlight
+  if (geom == "raster")
+    g <- g + ggplot2::geom_raster(ggplot2::aes(fill = x))
+
+  if (geom == "circle")
+    if (highlight) {
+      g <-
+        g + ggplot2::geom_point(ggplot2::aes(size = x,
+          fill = x > highlight),
+          color = "black",
+          pch = 21) + ggplot2::scale_fill_manual(values = c("white", "black")) +
+        ggplot2::guides(fill = FALSE, size = FALSE)
+    } else{
+      g <- g + ggplot2::geom_point(ggplot2::aes(size = x))
+    }
+
+  if (geom == "block")
+    if (highlight) {
+      g <-
+        g + ggplot2::geom_tile(
+          ggplot2::aes(
+            x = col,
+            y = row,
+            height = x / max(x) * .8,
+            fill = x > highlight
+          ),
+          width = .8,
+          color = "black"
+        ) +
+        ggplot2::scale_fill_manual(values = c("white", "black")) +
+        ggplot2::guides(fill = FALSE)
+    } else{
+      g <- g +
+        ggplot2::geom_tile(ggplot2::aes(height = x / max(x) * .9), width = .8)
+    }
+
+  # no highlight for line
+  if (geom == "line")
+    g <- g +
+    ggplot2::geom_line(ggplot2::aes(x = col, y = x, group = row)) +
+    # Note: facets display the lowest level first so we need to reverse them
+    ggplot2::facet_grid(rows = ggplot2::vars(stats::reorder(row, rev(as.integer(row))))) +
+    ggplot2::theme(
+      strip.text.y.right = ggplot2::element_text(angle = 0, color = "black"),
+      strip.background = ggplot2::element_blank()
+    )
+
+  if (geom == "bar")
+    if (highlight) {
+      g <- g +
+        ggplot2::geom_bar(ggplot2::aes(
+          x = col,
+          y = x,
+          group = row,
+          fill = x > highlight
+        ),
+          stat = "identity", color = "black", width = .8) +
+        # Note: facets display the lowest level first so we need to reverse them
+        ggplot2::facet_grid(rows = ggplot2::vars(stats::reorder(row, rev(as.integer(row))))) +
+        ggplot2::theme(
+          strip.text.y.right = ggplot2::element_text(angle = 0, color = "black"),
+          strip.background = ggplot2::element_blank()
+        ) +
+        ggplot2::scale_fill_manual(values = c("white", "black")) +
+        ggplot2::guides(fill = FALSE)
+    } else{
+      g <- g +
+        ggplot2::geom_bar(ggplot2::aes(x = col,
+          y = x,
+          group = row),
+          stat = "identity", width = .8) +
+        # Note: facets display the lowest level first so we need to reverse them
+        ggplot2::facet_grid(rows = ggplot2::vars(stats::reorder(row, rev(as.integer(row))))) +
+        ggplot2::theme(
+          strip.text.y.right = ggplot2::element_text(angle = 0, color = "black"),
+          strip.background = ggplot2::element_blank()
+        )
+    }
+
+  g
 }